Refactored code, created config header
