+++ include/geowars/game_components.hpp
