+++ src/core/game.cpp
@@ -15,93 +15,11 @@ namespace GWars
 {
 	Game::Game(const std::string& config)
 	{
-		// Verify config file can be located and parsed, if not, print an error message
-		// Error handling
-		try
-		{
-		std::ifstream f(config);
-		json game_config = json::parse(f);
-
 		init(game_config);
-		}
-
-		catch (const json::parse_error& e)
-		{
-			std::cout << "Message: Couldn't open config file for reading. \n"
-					<< "Error: " << e.what() << "Exception ID: " << e.id << "\n";
-			exit(-1);
-		}
-
 	}
 
 	void Game::init(const json& game_config)
 	{	
-		// Set window configurations
-		m_windowConfig.W = game_config["window"]["width"];
-		m_windowConfig.H = game_config["window"]["height"];
-		m_windowConfig.FR = game_config["window"]["frame_rate"];
-		m_windowConfig.UNK = game_config["window"]["fullscreen"];	
-
-		// Set on screen text strings
-		m_textConfig.WT = game_config["text"]["window_text"];
-		m_textConfig.ST = game_config["text"]["screen_text"];
-
-		// Set font configurations
-		m_fontConfig.F = game_config["font"]["file"];
-		m_fontConfig.S = game_config["font"]["size"];
-		m_fontConfig.R = game_config["font"]["color"].at(0);	
-		m_fontConfig.G = game_config["font"]["color"].at(1);
-		m_fontConfig.B = game_config["font"]["color"].at(2);
-
-		// Set player, enemy, and bullet configurations
-		for (auto& a : game_config["entity"].items())
-		{
-			EntityConfig temp;
-		
-			temp.T = a.key();
-			temp.SR = game_config["entity"][temp.T]["shape_radius"];
-			temp.CR = game_config["entity"][temp.T]["collision_radius"];
-			temp.SMin = game_config["entity"][temp.T]["speed"].at(0);
-			temp.SMax = game_config["entity"][temp.T]["speed"].at(1);
-			temp.FR = game_config["entity"][temp.T]["fill_color"].at(0);	
-			temp.FG = game_config["entity"][temp.T]["fill_color"].at(1);		
-			temp.FB = game_config["entity"][temp.T]["fill_color"].at(2);	
-			temp.OR = game_config["entity"][temp.T]["outline_color"].at(0);	
-			temp.OG = game_config["entity"][temp.T]["outline_color"].at(1);		
-			temp.OB = game_config["entity"][temp.T]["outline_color"].at(2);		
-			temp.OT = game_config["entity"][temp.T]["outline_thickness"];
-			temp.VMin = game_config["entity"][temp.T]["vertices"].at(0);		
-			temp.VMax = game_config["entity"][temp.T]["vertices"].at(1);
-			temp.SL = game_config["entity"][temp.T]["spawn_lifespan"];	
-			temp.SI = game_config["entity"][temp.T]["spawn_interval"];	
-			
-			if (temp.T == "player")
-				m_playerConfig = temp;
-			if (temp.T == "enemy")
-				m_enemyConfig = temp;
-			if (temp.T == "bullet")
-				m_bulletConfig = temp;
-		}
-
-		// Set asset configurations
-		m_terrainConfig.F = game_config["terrain"]["file"];
-		m_terrainConfig.N = game_config["terrain"]["name"];
-		m_terrainConfig.D = game_config["terrain"]["difficulty"];
-		
-		// Load and verify font can be loaded, if not, print an error message
-		if (!m_font.loadFromFile(m_fontConfig.F))
-		{
-			std::cerr << "Could not load font!\n";
-			exit(-1);
-		}
-		
-		// Load and verify terrain texture can be loaded, if not, print an error message
-		if (!m_terrain.loadFromFile(m_terrainConfig.F))
-		{
-			std::cerr << "Could not load terrain texture!\n";
-			exit(-1);
-		}
-
 		// Spawn the player
 		spawnPlayer();
 	}
